payment update v1

+++ app/src/main/java/com/example/ecoplay_front/model/CreatePaymentIntentRequest.kt
@@ -1,5 +1,5 @@
 package com.example.ecoplay_front.model
 
 data class CreatePaymentIntentRequest(
-    val totalC: Int
+    val totalC: Double
 )
